pyproject.toml
[tool.poetry]
name = "dbt-4-superset"
version = "0.3.4"
description = "Make dbt docs and Apache Superset talk to one another"
authors = ["Michal Kolacek <mkolacek@slido.com>", "AIBUNNY <bunny@theaibunny.com>"]
readme = "README.md"
repository = "https://github.com/aibunny/dbt-4-superset"

[tool.poetry.scripts]
smartcollect-lineage = "smartcollect_lineage.__init__:app"

[tool.poetry.dependencies]
python = "^3.8"
typer = "^0.4.0"
pathlib = "^1.0.1"
"ruamel.yaml" = "^0.17.17"
requests = "^2.26.0"
bs4 = "^0.0.1"
Markdown = "^3.3.6"
sqlfluff = ">=1.4.1,<3.0.0"
pytest = "^7.2.2"
python-dotenv = "^1.0.1"
pyyaml = "^6.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
